feat(active-active): Fail StartWorkflow request if cluster attribute doesn't exist (#7485)
<!-- Describe what has changed in this PR -->
**What changed?**
Fail StartWorkflow request if cluster attribute doesn't exist

<!-- Tell your future self why have you made these changes -->
**Why?**
We should avoid changing requests implicitly, and should fail and let
the customer decide how to handle the failure instead.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit test

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
This could break customers, but the feature is not available in open
source yet. Internally, we only have 1 customer onboarded, and the risk
of breaking that customer is very low.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**