fix(RangeCompleteHistoryTask): fix infinite loop when page size provided is <= 0 (#7472)
<!-- Describe what has changed in this PR -->
**What changed?**
Add a check to exit completion loop if pageSize is less or equal to 0.

<!-- Tell your future self why have you made these changes -->
**Why?**
When pageSize/batchSize is set to zero it's an unbounded request for
completion in sql dbs. It's not used by non sql dbs. That means that we
don't need to execute it again within the same request.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit test

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: fimanishi <fimanishi@gmail.com>