fix(jitter): allow input equal to 0 to be provided without panic (#7481)
<!-- Describe what has changed in this PR -->
**What changed?**
Check if input to Jit is zero and if it is, return the input as the
result

<!-- Tell your future self why have you made these changes -->
**Why?**
JitInt64 calls rand.Int63n(0) which panics because you can't get a
random number in the range [0,0). Same for JitFloat64.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit test

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Currently, if the value is 0 it panics. This address the panic and
allows the input to be zero.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: fimanishi <fimanishi@gmail.com>