refactor: [shard-distributor]Remove error logs from store level (#7492)
<!-- Describe what has changed in this PR -->
**What changed?**
Removing the "Store failed with internal error." and "Executor not
found." from the store level.

<!-- Tell your future self why have you made these changes -->
**Why?**
Since the error is returned in the call() method, returning the error
message is handled in the higher level.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests & local testing

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**
Cleaning up logs.

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>