feat(matching): use ShardProcessors instead of TaskListManagers (#7480)
<!-- Describe what has changed in this PR -->
**What changed?**
- Instead of using tasklistmanagers we start using the shardprocessor to
onboard matching to the shardDistributor.
- The shardProcessor factory is introduced and used to create new
shardProcessors.
- The executor client library is changed to have Start() function which
returns an error.

<!-- Tell your future self why have you made these changes -->
**Why?**
The executor library is using the concept of shardProcessor to monitor a
shard, in the case of matching a shardProcessor can be seen as a task
executor manager.
The change in the executor library is necessary because moving the logic
from the

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
tested locally and with unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
matching service could be affected since we are changing the way we
create new tasks.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: edigregorio <edigregorio@uber.com>