fix(shard-manager): fixed the ETCD integration test (#7502)
<!-- Describe what has changed in this PR -->
**What changed?**
Added an additional pattern to look for when checking if a package has
ETCD integration tests

<!-- Tell your future self why have you made these changes -->
**Why?**
A helper setting up the tests was matching the pattern, not the packages
that needed testing.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Tested locally that now unit tests fails if we break them 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Jakob Haahr Taankvist <jht@uber.com>