feat: adds cluster-attributes to start cli command (#7494)
<!-- Describe what has changed in this PR -->
**What changed?**

Adds the ability to specify cluster attribute and cluster scope in the
workflow start command. The goal is for customers who want to start
workflows in multiple regions using the active-active feature, to be
able to do so with manual cli invocations.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
tested locally for the start call, seems to work as expected. 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>