fix(shard-distributor-ex-client): handle concurrent assignment (#7500)
<!-- Describe what has changed in this PR -->
**What changed?**
instead of acquiring the lock when we want to assign the new state, we
are acquiring it before the heartbeat, to guarantee that we are going to
assign the result of the heartbeat just performed.


<!-- Tell your future self why have you made these changes -->
**Why?**
there are two situations that could be problematic:
- first it is not guaranteed that we will process the heartbeats in
order, so we could have stale info for some time. this is not a huge
problem because the heartbeat is retryied regularly so the information
will be updated at some point
- second this change is needed because during the migration we are
performing local assignment and in this case if it is overwritten by the
new heartbeat is really dangerous.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
local environment

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>