chore(shard-distirbutor): extend info to debug assignment conflicts (#7506)
<!-- Describe what has changed in this PR -->
**What changed?**
Extending the transaction with operation in the else branch to get the
state in case the comparison fails

<!-- Tell your future self why have you made these changes -->
**Why?**
we need to understand why there is a conflict when we assign the shards
but it is not possible because etcd is not showing which is the failing
comparison, so we need to be able to understand where the conflict is
generated

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests that cover this case actually reported the mismatch

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: edigregorio <edigregorio@uber.com>