refactor(shard-distributor): store shard stats under executor keys (#7507)
<!-- Describe what has changed in this PR -->
**What changed?**
- Statistics for shards are now stored under executor keys at
**\<prefix>/\<namespace>/executors/\<executorID>/statistics** instead of
under **/shards/.../statistics**. `GetState` is also updated to support
this.
- `DeleteShardStats` deletes shard stats by editing those executor maps
and committing a single txn.
- We no longer do batching as I don't believe we will reach etcd limits
anymore (stats under ~128 different executors that are stale doesn't
seem plausible)
- Delete now unused helpers and tests for shard keys
  
- The Subscribe logic was updated so changes to the statistics key are
treated like heartbeats/assigned_state (so they dont trigger
rebalances), and the etcd store tests were updated to use the new
executor keyed setup.

<!-- Tell your future self why have you made these changes -->
**Why?**
Reduce load to etcd

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit test and running the canary

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

- The etcd schema for shard statistics has changed, so any leftover
stats keys under shard prefix will be ignored by the new code.

- DeleteShardStats now issues all per-executor updates in a single
transaction. So in an extreme scenario with many executors in one
cleanup pass this could hit etcd’s op limit and leave some stale stats
until the next run. Consider if it's worth adding batching if this we
think this is ever possible.

- Concurrent AssignShard calls to the same executor now share a single
stats map without a CAS on the stats key, so shard telemetry (not
assignments) could be lost in a rare ocasion where there will be a race
which might affect future metrics or load-based decisions. But we made
this trade-off since it's telemetry, and we don't want it to cause
assignments to retry.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Andreas Holt <6665487+AndreasHolt@users.noreply.github.com>
Signed-off-by: Jakob Haahr Taankvist <jht@uber.com>
Co-authored-by: Jakob Haahr Taankvist <jht@uber.com>