fix: field mapping for list-failover-history should do graceful/force failover types  (#7522)
<!-- Describe what has changed in this PR -->
**What changed?**
This fixes a field mapping for failover type which I missed. While we
don't support graceful failover with the failover endpoint presently,
and may not decide to support it in the future, it's a valid type of
failover for now and we should include it's info in the history.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

tested by running locally and printing the output in JSON, where it's
rendered

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>