feat: making error-injection less annoying on startup (#7532)
<!-- Describe what has changed in this PR -->
**What changed?**

Error injection is a great, existing way to find bugs in the cadence
server side. Right now, however, it's annoying to use because the
probability of hitting a persistence error on startup is high enough and
the amount of start-time critical calls that will fail the service start
is so high that in practice the startup just fails to the point it's not
usable.

To be sure, there's value in testing the service startup, but most of
the interesting bugs I think we're looking for aren't in service
startup.

This:
- Adds a timeout of 30 seconds before starting to fail persistence calls

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>