fix: Cleanup needs to handle signal-with-start (#7540)
<!-- Describe what has changed in this PR -->
**What changed?**

We need to ensure the cleanup code doesn't affect signal-with-start API
calls. This flow is way too complicated and in need of a refactor, but
for now this just adds some additional defensiveness and a small amount
of additional cleanup to the GC flow during workflow start failure.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Tested by running locally with an XDC setup and injected errors,
although there's still probably some gaps in the amount of testing that
is needed. the PR is pretty defensive however, so seems less likely to
be a significant risk.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>