feat(cadence-matching): Onbord to use ShardDistributor executor client (#7533)
<!-- Describe what has changed in this PR -->
**What changed?**
- Matching is using the executor client to model the shards that the
instance own. Every time a new tasklist is created a new shard is also
added in the executor library, the shardID is calculated the same way it
is right now with the ringpop hasring, based on hashing of the tasklist
name.
- Matching is only working in local passthrough mode, which means no
communication with shard distributor is happening after this pr, and the
default behavior is maintained.
- Introcuced the concept of TTL for shards, if the tasklists are not
being used for a time which is aligning with tasklists liveness, than
the executor is removing the shards.

<!-- Tell your future self why have you made these changes -->
**Why?**
We want to onboard matching to use shard distributor for shard
balancing.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests and integration tests run locally and CI, plus deployed
matching instance in a staging environment.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Performance degradations of matching service, since the changes are
affecting the core logic of how we handle tasklists, adding additional
operations on the lock that controls all the tasklists.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>