fix(shard-distributor-canary): remove noisy log (#7572)
<!-- Describe what has changed in this PR -->
**What changed?**
Removing a log line that is too noisy

<!-- Tell your future self why have you made these changes -->
**Why?**
The size of the logs is increasing too much due to this new log line
added recently

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Checked in a deployed environment what was the log that is produced the
most

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
no risk, it is just a log line in a test service

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: edigregorio <edigregorio@uber.com>