fix: correct the detection of DB consistency error (#7573)
<!-- Describe what has changed in this PR -->
**What changed?**

A recent DB node failure tested this detection of Cassandra unavailable
errors in the wild and it didn't work as expected, it didn't identify a
failure (on Cassandra 4.1.10.2) that the nodes were not enough to form a
Quorum, despite the logs clearly indicating it.

Looking at the code in Cassandra Go client lib, it's not disambiguating
Quorum failures from other 5XX type errors. Maybe we should be looking
at all unavailability errors, but the major production issue we're
observing is that the loss of 2 nodes in Quorum (given RF 3 - the
industry default for this kind of thing) indicates a more series problem
and one that isn't merely a transient timeout. Something that requires
intervention.

### Testing 

I replicated a failure to catch this problem (it's just a bad string
match) using:
- docker-compose for 3 nodes using Cassandra 4.1.3
- RF 3 for the keyspaces
- a loss of two modes manually 
- Both LWT and normal writes (using consistency quorum)

---------

Signed-off-by: David Porter <david.porter@uber.com>